requirements.txt
torch
transformers
pandas
scikit-learn
tqdm
pyyaml
fastapi
uvicorn
